requirements.txt
python-binance>=1.0.17
pandas>=2.0.0
numpy>=1.24.0
rich>=13.0.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
